# frontend/openapi/backend-api.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "description": "Formato común y seguro de los errores controlados de la API.",
        "properties": {
          "code": {
            "description": "Código público y estable del error.",
            "example": "VALIDATION_ERROR",
            "type": "string"
          },
          "correlationId": {
            "description": "Identificador de correlación para soporte y trazabilidad.",
            "example": "7a5f3f75-3bd2-4c47-90fc-6cfc79f1ec2d",
            "type": "string"
          },
          "message": {
            "description": "Mensaje comprensible sin detalles internos.",
            "example": "La solicitud contiene datos inválidos.",
            "type": "string"
          },
          "path": {
            "description": "Ruta de la solicitud que produjo el error.",
            "example": "/api/v1/revocation-requests",
            "type": "string"
          },
          "timestamp": {
            "description": "Fecha y hora UTC del error.",
            "example": "2026-07-20T18:30:00Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "code",
          "correlationId",
          "message",
          "path",
          "timestamp"
        ],
        "type": "object"
      },
...
      "CurrentSession": {
        "description": "Contexto seguro de la operación activa y el siguiente paso autorizado.",
        "properties": {
          "dni": {
            "description": "DNI completo mostrado únicamente dentro de la sesión autenticada.",
            "pattern": "^[0-9]{8}$",
            "type": "string"
          },
          "nextStep": {
            "enum": [
              "IDENTITY_VERIFICATION",
              "DIGITAL_CREDENTIAL_SELECTION",
              "CONFIRMATION",
              "RECEIPT"
            ],
            "type": "string"
          },
          "requestId": {
            "description": "Identificador técnico de solicitud.",
            "format": "int64",
            "type": "integer"
          },
          "requestStatus": {
            "description": "Estado controlado de la solicitud.",
            "type": "string"
          },
          "sessionId": {
            "description": "Identificador técnico de sesión.",
            "format": "int64",
            "type": "integer"
          },
          "sessionStatus": {
            "enum": [
              "PENDING_IDENTITY",
              "IDENTITY_VERIFIED"
            ],
            "type": "string"
          }
        },
        "required": [
          "dni",
          "nextStep",
          "requestId",
          "requestStatus",
          "sessionId",
          "sessionStatus"
        ],
        "type": "object"
      },
...
      "RevocationRequestResponse": {
        "description": "Resultado normalizado del inicio de una nueva solicitud de revocación.",
        "properties": {
          "availabilityResult": {
            "description": "Resultado normalizado de la consulta inicial de existencia. No representa una lista detallada.",
            "enum": [
              "AVAILABLE",
              "NOT_AVAILABLE",
              "INCONCLUSIVE",
              "UNAVAILABLE",
              "ERROR"
            ],
            "example": "AVAILABLE",
            "type": "string"
          },
          "canContinue": {
            "description": "Indica si el backend autoriza continuar al siguiente paso.",
            "example": true,
            "type": "boolean"
          },
          "maskedDni": {
            "description": "DNI enmascarado para presentación segura.",
            "example": "******01",
            "type": "string"
          },
          "nextStep": {
            "description": "Siguiente paso autorizado. Es nulo cuando no se permite continuar.",
            "enum": [
              "IDENTITY_VERIFICATION"
            ],
            "example": "IDENTITY_VERIFICATION",
            "type": [
              "string",
              "null"
            ]
          },
          "requestId": {
            "description": "Identificador interno de la solicitud. No funciona como credencial ni autorización.",
            "example": 125,
            "format": "int64",
            "type": "integer"
          },
          "requestStatus": {
            "description": "Estado actual persistido de la solicitud.",
            "enum": [
              "STARTED",
              "CHECKING_AVAILABILITY",
              "NO_DIGITAL_CREDENTIALS_AVAILABLE",
              "PENDING_IDENTITY_VERIFICATION",
              "IDENTITY_VERIFIED",
              "AUTHENTICATED_PENDING_DIGITAL_CREDENTIAL_LIST",
              "CHECKING_DIGITAL_CREDENTIAL_LIST",
              "DIGITAL_CREDENTIALS_AVAILABLE",
              "DIGITAL_CREDENTIALS_SELECTED",
              "REASON_REGISTERED",
              "PENDING_CONFIRMATION",
              "CONFIRMED",
              "REVOCATION_IN_PROGRESS",
              "REVOCATION_SUCCEEDED",
              "REVOCATION_FAILED",
              "REVOCATION_OUTCOME_UNKNOWN",
              "COMPLETED",
              "FAILED",
              "OUTCOME_UNKNOWN",
              "RECEIPT_AVAILABLE",
              "ABANDONED"
            ],
            "example": "PENDING_IDENTITY_VERIFICATION",
            "type": "string"
          }
        },
        "required": [
          "availabilityResult",
          "canContinue",
          "maskedDni",
          "nextStep",
          "requestId",
          "requestStatus"
        ],
        "type": "object"
      },
...
      "StartRevocationRequest": {
        "description": "Datos requeridos para iniciar una nueva solicitud y consultar si existen credenciales disponibles.",
        "properties": {
          "dni": {
            "description": "Número de DNI del ciudadano. Debe contener exactamente ocho dígitos ASCII. Por privacidad, la documentación no incluye un DNI completo de ejemplo.",
            "maxLength": 8,
            "minLength": 8,
            "pattern": "[0-9]{8}",
            "type": "string"
          },
          "recaptchaToken": {
            "description": "Evidencia efímera de Google reCAPTCHA v2 Checkbox. No se almacena ni se devuelve.",
            "maxLength": 4096,
            "minLength": 0,
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "dni",
          "recaptchaToken"
        ],
        "type": "object"
      },
      "SystemStatusResponse": {
        "description": "Estado técnico del backend y su conexión con MySQL",
        "properties": {
          "database": {
            "description": "Disponibilidad de la conexión comprobada con MySQL.",
            "example": "UP",
            "type": "string"
          },
          "status": {
            "description": "Estado agregado del backend.",
            "example": "UP",
            "type": "string"
          },
          "timestamp": {
            "description": "Fecha y hora UTC de la comprobación.",
            "example": "2026-07-20T18:30:00Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "database",
          "status",
          "timestamp"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "FlowRefreshCookie": {
        "description": "Refresh JWT rotatorio, transportado exclusivamente mediante cookie HttpOnly.",
        "in": "cookie",
        "name": "revocacion_refresh",
        "type": "apiKey"
      },
      "FlowSessionCookie": {
        "description": "Access JWT de la sesión transaccional, transportado exclusivamente mediante cookie HttpOnly.",
        "in": "cookie",
        "name": "revocacion_access",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "RENIEC"
    },
    "description": "API institucional para iniciar el flujo ciudadano. La consulta pública exige Google reCAPTCHA v2 Checkbox y la autenticación usa ID Perú con continuidad temporal HttpOnly.",
    "title": "API de revocación de credenciales digitales",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/revocation-requests": {
      "post": {
        "description": "Valida el DNI y Google reCAPTCHA v2 Checkbox antes de crear una solicitud y consultar únicamente si existe al menos una credencial disponible para revocar. No obtiene una lista, cantidad, número de orden, fecha de creación ni UUID; tampoco reabre solicitudes anteriores ni expone el DNI completo o la evidencia anti-bot.",
        "operationId": "initiateRevocationRequest",
        "parameters": [
          {
            "description": "Identificador opcional de correlación. Debe tener entre 1 y 64 caracteres ASCII válidos.",
            "in": "header",
            "name": "X-Correlation-ID",
            "schema": {
              "maxLength": 64,
              "pattern": "[A-Za-z0-9][A-Za-z0-9._-]{0,63}",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRevocationRequest"
              }
            }
          },
          "description": "DNI y evidencia efímera reCAPTCHA que se validan en el backend antes de crear una solicitud.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationRequestResponse"
                }
              }
            },
            "description": "Resultado normalizado de existencia de credenciales",
            "headers": {
              "X-Correlation-ID": {
                "description": "Identificador de correlación",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "DNI, JSON, cuerpo o evidencia reCAPTCHA inválida: VALIDATION_ERROR, RECAPTCHA_REQUIRED, RECAPTCHA_REJECTED o RECAPTCHA_EXPIRED_OR_DUPLICATE",
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto controlado: AVAILABILITY_CHECK_IN_PROGRESS, REVOCATION_REQUEST_IN_PROGRESS o CONCURRENT_REQUEST",
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Tipo de contenido no admitido",
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno controlado",
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "RECAPTCHA_INVALID_RESPONSE o error controlado del proveedor de disponibilidad",
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "RECAPTCHA_UNAVAILABLE o servicio de disponibilidad no disponible",
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "RECAPTCHA_TIMEOUT o tiempo de espera agotado en disponibilidad",
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          }
        },
        "summary": "Inicia una solicitud y consulta si existen credenciales disponibles",
        "tags": [
          "Solicitudes de revocación"
        ]
      }
    },
...
    "/api/v1/session/current": {
      "get": {
        "operationId": "getCurrentFlowSession",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentSession"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Consulta la sesión y el paso actualmente autorizado",
        "tags": [
          "Sesión del flujo"
        ]
      }
    },
    "/api/v1/session/logout": {
      "post": {
        "operationId": "logoutFlowSession",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "FlowSessionCookie": []
          }
        ],
        "summary": "Cierra la sesión y abandona la operación activa reversible",
        "tags": [
          "Sesión del flujo"
        ]
      }
    },
    "/api/v1/session/refresh": {
      "post": {
        "operationId": "refreshFlowSession",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "FlowRefreshCookie": []
          }
        ],
        "summary": "Rota el refresh token y actualiza el access token",
        "tags": [
          "Sesión del flujo"
        ]
      }
    },
    "/api/v1/system/status": {
      "get": {
        "description": "Ejecuta una comprobación ligera y actual de MySQL. No devuelve credenciales, coordenadas de conexión ni detalles internos.",
        "operationId": "getSystemStatus",
        "parameters": [
          {
            "description": "Identificador opcional de correlación. Debe tener entre 1 y 64 caracteres ASCII válidos.",
            "in": "header",
            "name": "X-Correlation-ID",
            "schema": {
              "maxLength": 64,
              "pattern": "[A-Za-z0-9][A-Za-z0-9._-]{0,63}",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusResponse"
                }
              }
            },
            "description": "Backend y base de datos disponibles",
            "headers": {
              "X-Correlation-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno controlado",
            "headers": {
              "X-Correlation-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "MySQL no está disponible",
            "headers": {
              "X-Correlation-ID": {
                "description": "Identificador de correlación de la solicitud",
                "schema": {
                  "type": "string"
                },
                "style": "simple"
              }
            }
          }
        },
        "summary": "Comprueba la disponibilidad del backend y MySQL",
        "tags": [
          "Estado técnico"
        ]
      }
    }
  },
...
}